If in a cargo package, use cargo verifast
The extension now looks for a Cargo.toml file in the active file's directory or any ancestor, and if it finds one, it runs 'cargo verifast' instead of 'verifast'. This is useful for specifying the root source file, the Rust edition, and other rustc flags, and in case the program has dependencies.

It adds the directory of the configured VeriFast executable to the front of the PATH for running 'cargo verifast'.

Also:
- The extension now clears the old symbolic execution trace and any old inline problem view before running VeriFast.